# Range deletion microoptimizations